feat(date): add date feature
implement input, output, parser, validate method

+++ src/main/java/christmas/constant/UtecoMenu.java
@@ -1,27 +1,31 @@
 package christmas.constant;
 
 public enum UtecoMenu {
+    // 애피타이저
     BUTTON_MUSHROOM_SOUP("양송이수프", 6_000),
     TAPAS("타파스", 5_500),
     CAESAR_SALAD("시저샐러드", 8_000),
 
+    // 메인
     T_BONE_STEAK("티본스테이크", 55_000),
     BARBECUE_RIBS("바비큐립", 54_000),
     SEAFOOD_PASTA("해산물파스타", 35_000),
     XMAS_PASTA("크리스마스파스타", 25_000),
 
+    // 디저트
     CHOCOLATE_CAKE("초코케이크", 15_000),
     ICE_CREAM("아이스크림", 5_000),
 
+    // 음료
     ZERO_COKE("제로콜라", 3_000),
     RED_WINE("레드와인", 60_000),
     CHAMPAGNE("샴페인", 25_000),
     ;
 
     private final String menuName;
-    private final int price;
+    private final long price;
 
-    UtecoMenu(String menuName, int price) {
+    UtecoMenu(String menuName, long price) {
         this.menuName = menuName;
         this.price = price;
     }
@@ -30,7 +34,7 @@ public enum UtecoMenu {
         return menuName;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 }
